Move StandardIdentifiers to Detectors module
Why these changes are being introduced:

Since we created the `StandardIdentifiers` model, we started
namespacing our detection models.

Relevant ticket(s):

Tangentially related to [TCO-50](https://mitlibraries.atlassian.net/browse/TCO-50)

How this addresses that need:

This adds the `Detectors` namespace to `StandardIdentifiers` to
match our other detection models (`Journal` and `SuggestedResource`).

Side effects of this change:

Some cascading changes were made in Metrics::Algorithms and
the SearchEventType GraphQL.
